+++ Sheeeeeeeeet/Items/Buttons/ActionSheetCancelButton.swift
@@ -11,7 +11,7 @@
  Cancel buttons have no special behavior, but can be used in
  sheets where a user applies changes by tapping an OK button.
  
- The value of a cancel button value is `nil`.
+ The value of a cancel button is `ButtonType.cancel`.
  
  */
 
@@ -23,7 +23,7 @@ open class ActionSheetCancelButton: ActionSheetButton {
     // MARK: - Initialization
     
     public init(title: String) {
-        super.init(title: title, value: nil)
+        super.init(title: title, type: .cancel)
     }