+++ Sheeeeeeeeet/Items/Buttons/ActionSheetOkButton.swift
@@ -11,7 +11,7 @@
  OK buttons have no special behavior, but can be used when a
  user should apply action sheet changes by tapping an button.
  
- The value of an OK button is `true`.
+ The value of an OK button is `ButtonType.ok`.
  
  */
 
@@ -23,7 +23,7 @@ open class ActionSheetOkButton: ActionSheetButton {
     // MARK: - Initialization
     
     public init(title: String) {
-        super.init(title: title, value: true)
+        super.init(title: title, type: .ok)
     }