+++ Sheeeeeeeeet/Items/Titles/ActionSheetSectionMargin.swift
@@ -26,11 +26,6 @@ open class ActionSheetSectionMargin: ActionSheetItem {
     }
     
     
-    // MARK: - Properties
-    
-    open override var itemType: ItemType { return .title }
-    
-    
     // MARK: - Functions
     
     open override func applyAppearance(_ appearance: ActionSheetAppearance) {