+++ SheeeeeeeeetTests/Items/Buttons/ActionSheetCancelButtonTests.swift
@@ -20,7 +20,11 @@ class ActionSheetCancelButtonTests: ActionSheetItemTests {
             
             it("applies provided values") {
                 expect(item.title).to(equal("foo"))
-                expect(item.value).to(beNil())
+            }
+            
+            it("is correctly setup") {
+                expect(item.value as? ActionSheetButton.ButtonType).to(equal(.cancel))
+                expect(item.isCancelButton).to(beTrue())
             }
         }